adicionado arquivo com funções utilizadas para todos os arquivos de teste
