+++ desapeg/tests/test_send_form.py
@@ -3,14 +3,7 @@ import io
 from PIL import Image
 from desapeg.app import app
 from desapeg.extensions import db
-
-# Função auxiliar para criar uma imagem válida em memória
-def create_test_image():
-    img = Image.new('RGB', (10, 10), color='red')
-    img_bytes = io.BytesIO()
-    img.save(img_bytes, format='JPEG')
-    img_bytes.seek(0)
-    return img_bytes
+from desapeg.tests.test_utils import create_test_image
 
 class ProductFormTestCase(unittest.TestCase):
