+++ desapeg/tests/test_utils.py
@@ -1,0 +1,10 @@
+import io
+from PIL import Image
+
+# Função auxiliar para criar uma imagem válida em memória
+def create_test_image():
+    img = Image.new('RGB', (10, 10), color='red')
+    img_bytes = io.BytesIO()
+    img.save(img_bytes, format='JPEG')
+    img_bytes.seek(0)
+    return img_bytes
